% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzar.R
\name{bzar.get_data}
\alias{bzar.get_data}
\title{Get Time Series Data}
\usage{
bzar.get_data(
  station_type = "",
  station_code = "",
  data_set = "",
  from_ts = "",
  to_ts = "",
  period_hint = 1
)
}
\arguments{
\item{station_type}{The station type.}

\item{station_code}{The station code.}

\item{data_set}{The data set.}

\item{from_ts}{Start (>=) of the interval as string of the form yyyy-MM-dd[T[HH][:mm][:ss][.SSS]][Z], where Z is the timezone (e.g. +0200).}

\item{to_ts}{End (<) of the interval as string of the form yyyy-MM-dd[T[HH][:mm][:ss][.SSS]][Z], where Z is the timezone (e.g. +0200).}

\item{period_hint}{The preferred sample period in seconds (defaults to the smallest available).}
}
\value{
A data frame with the requested time series in the half-open interval [from_ts, to_ts).
}
\description{
Retrieve the time series data corresponding to the given station type,
station code, data set and time interval. Optionally specify the desired
sample period (defaults to smallest available).
}
