% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzar.R
\name{bzar.get_data_sets}
\alias{bzar.get_data_sets}
\title{Get Data Sets}
\usage{
bzar.get_data_sets(station_type = "", station_code = "")
}
\arguments{
\item{station_type}{The station type.}

\item{station_code}{The station code.}
}
\value{
A data frame with the data set name and units.
}
\description{
Given a station type and a station code, return a data frame with the
available data sets. Use bzar.get_station_types() to get a vector of known
station types. Use bzar.get_stations() to get a data frame of known
station codes.
}
