% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzar.R
\name{bzar.get_station_types}
\alias{bzar.get_station_types}
\title{Get Station Types}
\usage{
bzar.get_station_types()
}
\value{
A vector with the station types names.
}
\description{
Get the available station types.
}
