% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bzar.R
\name{bzar.get_stations}
\alias{bzar.get_stations}
\title{Get Stations}
\usage{
bzar.get_stations(station_type = "")
}
\arguments{
\item{station_type}{The station type.}
}
\value{
A data frame with the station names and codes.
}
\description{
Given a station type, return a data frame with the available stations.
Use bzar.get_station_types() to get a vector of known station types.
}
